% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{estBeta}
\alias{estBeta}
\title{Estimate the regression coefficients}
\usage{
estBeta(iter_res, burnin)
}
\arguments{
\item{iter_res}{output from BMVS()}

\item{burnin}{burn-in iterations}
}
\value{
estimate regression coefficients
}
\description{
Regression coefficients
}
\examples{
estBeta(BMVS_res, burnin = 5000)
}
