% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMVS.R
\name{iter}
\alias{iter}
\title{Run one iteration of the MCMC chain}
\usage{
iter(
  Y,
  X,
  n,
  p,
  q,
  par_list,
  tao0,
  tao1,
  qn,
  beta_vec,
  z_vec,
  z_prob_vec,
  sigma2_vec,
  invY_vec,
  alpha1,
  alpha2,
  vk,
  deltak,
  cur
)
}
\arguments{
\item{Y}{input response variables}

\item{X}{input predictors}

\item{n}{sample size}

\item{p}{the number of predictors}

\item{q}{the number of response variables}
}
\value{
The results of one MCMC iteration
}
\description{
This is an internal function called by BMVS()
}
