% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select an optimal model}
\usage{
select(iter_res, Y, X, burnin)
}
\arguments{
\item{iter_res}{output from BMVS()}

\item{Y}{input response variables}

\item{X}{input predictors}

\item{burnin}{burn-in iterations}
}
\value{
indices of selected predictors
}
\description{
Select an optimal model using BIC.
The inputs are the output from BMVS() function call.
}
\examples{
BMVS::select(BMVS_res, Y, X, burnin = 5000)
}
